requirements.txt
flask
Flask-Session
requests
beautifulsoup4
gunicorn
pymongo[srv]
python-dotenv
sklearn
redis
pillow
schedule
nltk
rake-nltk
imagehash